Initial commit.
This represents the state of the site as at 2013-05-08, and is content created by a number of people and reformatted for Statamic over recent months.

This is not the entire file set needed to run Statamic, nor does it contain some of the files for the theme that under a restricted license. Contact the site owner if you need to use those in order to setup a local copy to preview your edits.